+++ src/core.rs
@@ -1,3 +1,4 @@
+pub mod open_simplex;
 pub mod perlin;
 pub mod perlin_surflet;
 pub mod simplex;
